Ok, many things.
1) I moved the google map view to the url explorer/map,

2) I connected django with a mysql server on the localhost.

3) I created data models in explorer/models.py for the medicare
database. This created a bunch of tables in mysql

4) I used the iPython notebook database_loader.ipynb to create
a dictionary {procedure_code:procedure_descriptor}. Then I loaded
the entries of the dictionary in the corresponding table in mysql
using the django database api.

5) I created a view associated with the url explorer/ that allows
instant search through the procedure dictionary

Next: I want to show the coordinates of the providers on the
google map.