I created a complete model Provider, then run migrations. I temporarily removed the model Procedure, because it was not complete, it was missing many fields and I don't fill it now.
I added a script to database_loader.ipynb to load the Provider table
with actual data. It uses geopy to find latitude and longitude. It's
really slow. Filling the table in full will probably take a few hours,
for now I just put 100 providers in.

Once the data was loaded, I added a couple of lines to the google_map_data
view that display the provider locations on the map.

Next, I want to display on the same page as the map the list of providers
within the map viewport.